+++ lib/src/modules/moyasar/views/uni_pay_moyasar_view.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:moyasar/moyasar.dart';
 import 'package:uni_pay/src/utils/extension/size_extension.dart';
 import 'package:uni_pay/src/views/design_system.dart';
-import 'package:uni_pay/src/views/uni_apple_pay.dart';
 
 import '../../../constant/uni_text.dart';
 import '../../../core/keys/api_keys.dart';
+import '../../applepay/uni_apple_pay.dart';
 import '../core/services/uni_moyasar.dart';
 
 class UniPayCard extends StatefulWidget {
@@ -18,9 +18,9 @@ class UniPayCard extends StatefulWidget {
 class _UniPayCardState extends State<UniPayCard> {
   @override
   Widget build(BuildContext context) {
-    return UniPayScaffold(
-      title: UniPayText.payWithCard,
-      builder: (_) => ListView(
+    return Scaffold(
+      appBar: UniPayDesignSystem.appBar(title: UniPayText.payWithCard),
+      body: ListView(
         padding: EdgeInsets.all(20.rSp),
         children: [
           const UniPayApplePay(),