+++ lib/src/modules/tamara/core/models/t_checkout.dart
@@ -1,14 +1,20 @@
+import 'package:uni_pay/src/constant/uni_text.dart';
+
 class TamaraCheckoutData {
+  late bool isSuccess;
   late String orderId;
   late String checkoutId;
   late String checkoutUrl;
   late String status;
+  String errorMessage = UniPayText.somethingWentWrong;
+  dynamic errors;
 
   TamaraCheckoutData({
-    required this.orderId,
-    required this.checkoutId,
-    required this.checkoutUrl,
-    required this.status,
+    this.isSuccess = false,
+    this.orderId = "",
+    this.checkoutId = "",
+    this.checkoutUrl = "",
+    this.status = "",
   });
 
   TamaraCheckoutData.fromJson(Map<String, dynamic> json) {
@@ -16,14 +22,18 @@ class TamaraCheckoutData {
     checkoutId = json['checkout_id'] ?? "";
     checkoutUrl = json['checkout_url'] ?? "";
     status = json['status'] ?? "";
+    isSuccess = checkoutUrl.isNotEmpty && orderId.isNotEmpty;
   }
 
   Map<String, dynamic> toJson() {
     final data = <String, dynamic>{};
+    data['isSuccess'] = isSuccess;
     data['order_id'] = orderId;
     data['checkout_id'] = checkoutId;
     data['checkout_url'] = checkoutUrl;
     data['status'] = status;
+    data['errorMessage'] = errorMessage;
+
     return data;
   }
 }