+++ lib/src/uni_pay.dart
@@ -1,9 +1,0 @@
-class UniPay {
-  UniPay._();
-
-  ///* Initialize UniPay all payment methods
-  static Future initUniPay() async {}
-
-  ///* Tamara checkout
-  // static Widget tamara() {}
-}