+++ lib/src/views/widgets/uni_pay_view_handler.dart
@@ -1,11 +1,12 @@
 import 'package:flutter/material.dart';
 import 'package:uni_pay/src/utils/extension.dart';
 
-import '../../uni_pay.dart';
-import '../modules/moyasar/views/uni_pay_moyasar_view.dart';
-import '../utils/extension/size_extension.dart';
-import 'design_system.dart';
-import 'uni_pay_all_view.dart';
+import '../../../uni_pay.dart';
+import '../../modules/moyasar/views/uni_pay_moyasar_view.dart';
+import '../../modules/tamara/views/tamara_pay_view.dart';
+import '../../utils/extension/size_extension.dart';
+import '../design_system.dart';
+import '../uni_pay_all_view.dart';
 
 class UniPayViewHandler extends StatefulWidget {
   const UniPayViewHandler({Key? key}) : super(key: key);
@@ -29,9 +30,7 @@ class _UniPayViewHandlerState extends State<UniPayViewHandler> {
   Future _viewHandler() async {
     ScreenSizes.init(context);
 
-    uniPayProivder.initUniPay();
-
-    final uniPayData = uniPayProivder.uniPayData;
+    final uniPayData = UniPayControllers.uniPayData;
 
     ///* Check for Tamara checkout
     if (uniPayData.credentials.paymentMethods.length == 1 &&
@@ -39,8 +38,8 @@ class _UniPayViewHandlerState extends State<UniPayViewHandler> {
       context.uniPushReplacement(const UniPayTamara());
     } else if ((uniPayData.credentials.paymentMethods.length == 1 &&
             uniPayData.credentials.paymentMethods.first.isMoyasar) ||
-        uniPayData.credentials.paymentMethods.length == 2 &&
-            uniPayData.credentials.paymentMethods.isMoyasarGateway) {
+        (uniPayData.credentials.paymentMethods.length == 2 &&
+            !uniPayData.credentials.paymentMethods.isTamaraGateway)) {
       context.uniPushReplacement(const UniPayCard());
     } else {
       context.uniPushReplacement(const UniPayGatewayView());